UI_AUTOMATION_TEST/requirements.txt
robotframework==6.0.2
playwright==1.48.0